Build complete archive of all four previous conference sites
Convert from a Jeckyl site to a static archive of all previous sites.
Updates to theme wound up being applied to all previous years
sites, causing the state and some information to be lost. Additionally,
the 2021 site was served from a different location using WordPress.

This patch takes all of the sites, as they were at the time of the
event, and makes a static archive of them. Some indexing changes
were made, but this update should capture the state of all previous
events and make way for new event management for 2022 and going
forward.

The new home for the conference is https://www.tvmcon.org